+++ .env.example
@@ -1,0 +1,2 @@
+CLOUDFLARE_API_TOKEN=
+MONGODB_URI=
