+++ .github/workflows/SST - Development.yml
@@ -33,6 +33,6 @@ jobs:
         with:
           role-to-assume: ${{ vars.ROLE_TO_ASSUME }}
           aws-region: ap-southeast-1
-      - name: Deploy app
-        run: |
-          pnpm install && pnpm sst deploy --stage dev
+#      - name: Deploy app
+#        run: |
+#          pnpm install && pnpm sst deploy --stage dev
