--- .vscode/launch.json
@@ -1,15 +1,0 @@
-{
-  "version": "0.2.0",
-  "configurations": [
-    {
-      "name": "Debug SST Start",
-      "type": "node",
-      "request": "launch",
-      "runtimeExecutable": "${workspaceRoot}/node_modules/.bin/sst",
-      "runtimeArgs": ["start", "--increase-timeout"],
-      "console": "integratedTerminal",
-      "skipFiles": ["<node_internals>/**"],
-      "env": {}
-    }
-  ]
-}
