--- .vscode/settings.json
@@ -1,5 +1,0 @@
-{
-  "search.exclude": {
-    "**/.sst": true
-  }
-}
