--- cdk.context.json
@@ -1,6 +1,0 @@
-{
-  "hosted-zone:account=173188516302:domainName=sgmotortrends.com:region=ap-southeast-1": {
-    "Id": "/hostedzone/Z028295527IED1A8GMZIH",
-    "Name": "sgmotortrends.com."
-  }
-}
