+++ env.d.ts
@@ -1,0 +1,6 @@
+declare namespace NodeJS {
+  export interface ProcessEnv {
+    CLOUDFLARE_API_TOKEN: string;
+    MONGODB_URI: string;
+  }
+}
