--- packages/core/package.json
@@ -1,17 +1,0 @@
-{
-  "name": "@lta-cars-dataset/core",
-  "version": "0.0.0",
-  "type": "module",
-  "scripts": {
-    "test": "sst bind vitest",
-    "typecheck": "tsc -noEmit"
-  },
-  "devDependencies": {
-    "vitest": "^0.34.6",
-    "@types/node": "^20.8.10",
-    "sst": "^2.34.3"
-  },
-  "dependencies": {
-    "zod": "^3.22.4"
-  }
-}
