--- packages/core/tsconfig.json
@@ -1,7 +1,0 @@
-{
-  "extends": "@tsconfig/node18/tsconfig.json",
-  "compilerOptions": {
-    "module": "esnext",
-    "moduleResolution": "node"
-  }
-}
