--- packages/functions/env.d.ts
@@ -1,7 +1,0 @@
-declare global {
-  namespace NodeJS {
-    interface ProcessEnv {
-      MONGODB_URI: string;
-    }
-  }
-}
