--- packages/functions/package.json
@@ -1,15 +1,0 @@
-{
-  "name": "@lta-cars-dataset/functions",
-  "version": "0.0.0",
-  "type": "module",
-  "scripts": {
-    "test": "sst bind vitest",
-    "typecheck": "tsc -noEmit"
-  },
-  "devDependencies": {
-    "@types/node": "^20.8.10",
-    "@types/aws-lambda": "^8.10.125",
-    "vitest": "^0.34.6",
-    "sst": "^2.34.3"
-  }
-}
