--- packages/functions/sst-env.d.ts
@@ -1,1 +1,0 @@
-/// <reference path="../../.sst/types/index.ts" />
