--- packages/functions/tsconfig.json
@@ -1,11 +1,0 @@
-{
-  "extends": "@tsconfig/node18/tsconfig.json",
-  "compilerOptions": {
-    "module": "esnext",
-    "moduleResolution": "node",
-    "baseUrl": ".",
-    "paths": {
-      "@lta-cars-dataset/core/*": ["../core/src/*"]
-    }
-  }
-}
