+++ src/config/db.ts
@@ -1,5 +1,4 @@
 import * as mongodb from "mongodb";
-import { Config } from "sst/node/config";
 
 const MongoClient = mongodb.MongoClient;
 
@@ -10,7 +9,7 @@ const connectToDatabase = async (): Promise<mongodb.Db> => {
     return cachedDb;
   }
 
-  const client = await MongoClient.connect(Config.MONGODB_URI);
+  const client = await MongoClient.connect(process.env.MONGODB_URI);
   cachedDb = await client.db("main");
 
   return cachedDb;
