+++ tsconfig.json
@@ -1,8 +1,6 @@
 {
-  "extends": "@tsconfig/node18/tsconfig.json",
-  "exclude": ["packages"],
   "compilerOptions": {
-    "module": "esnext",
-    "moduleResolution": "node"
+    "lib": ["esnext"],
+    "types": ["@cloudflare/workers-types"]
   }
 }
